+++ lib/hbase-jruby/admin.rb
@@ -1,8 +1,21 @@
 class HBase
 # @private
 module Admin
+  # Wait until the asynchronous alteration finishes
+  # @yield [progress, total]
+  # @yieldparam [Fixnum] progress Number of regions updated
+  # @yieldparam [Fixnum] total Total number of regions
+  # @return [nil]
+  def wait_async &block
+    with_admin do |admin|
+      wait_async_admin(admin, &block)
+    end
+  end
+
 private
   def with_admin
+    check_closed
+
     begin
       admin = HBaseAdmin.new(@config)
       yield admin
@@ -19,7 +32,7 @@ private
       total = pair.getSecond
 
       if block && yet != prev_yet
-        block.call (total - yet), total
+        block.call(total - yet, total)
         prev_yet = yet
       end